+++ DRN.Framework.Utils/Http/InternalRequest.cs
@@ -1,6 +1,7 @@
 using DRN.Framework.SharedKernel.Json;
 using DRN.Framework.Utils.DependencyInjection.Attributes;
 using DRN.Framework.Utils.Settings;
+using Flurl;
 using Flurl.Http;
 using Flurl.Http.Configuration;
 
@@ -41,7 +42,7 @@ public interface IInternalRequest
 }
 
 [Singleton<IInternalRequest>]
-public class InternalRequest(IAppSettings appSettings) : IInternalRequest
+public class InternalRequest(IAppSettings appSettings, HttpMessageHandler? httpMessageHandler) : IInternalRequest
 {
     private const string Http = "http://";
     private const string Https = "https://";
@@ -49,6 +50,11 @@ public class InternalRequest(IAppSettings appSettings) : IInternalRequest
     private static readonly DefaultJsonSerializer JsonSerializer = new(JsonConventions.DefaultOptions);
     private readonly string _httpVersionDefault = new Version(appSettings.Features.InternalRequestHttpVersion).ToString();
     private readonly bool _securityDefault = appSettings.Features.InternalRequestProtocol.StartsWith("https", StringComparison.OrdinalIgnoreCase);
+    private readonly IFlurlClient? _flurlClient = httpMessageHandler != null ? new FlurlClient(new HttpClient(httpMessageHandler, disposeHandler: false)) : null;
+
+    public InternalRequest(IAppSettings appSettings) : this(appSettings, null)
+    {
+    }
 
     public IFlurlRequest For(string service) => For(service, _securityDefault, _httpVersionDefault);
     public IFlurlRequest For(string service, bool secure) => For(service, secure, _httpVersionDefault);
@@ -59,17 +65,10 @@ public class InternalRequest(IAppSettings appSettings) : IInternalRequest
     /// <param name="httpVersion">1.1 or 2.0</param>
     public IFlurlRequest For(string service, bool secure, Version httpVersion) => For(service, secure, httpVersion.ToString());
 
-    private static IFlurlRequest For(string service, bool secure, string httpVersion)
+    private IFlurlRequest For(string service, bool secure, string httpVersion)
     {
         var protocol = secure ? Https : Http;
-        var flurlRequest = $"{protocol}{service}".WithSettings(x =>
-        {
-            x.HttpVersion = httpVersion;
-            x.JsonSerializer = JsonSerializer;
-        });
-
-        flurlRequest.BeforeCall(x => x.HttpRequestMessage.VersionPolicy = HttpVersionPolicy.RequestVersionExact);
-
-        return flurlRequest;
+        var url = $"{protocol}{service}";
+        return FlurlRequestFactory.Create(_flurlClient, url, httpVersion, JsonSerializer);
     }
 }