+++ DRN.Nexus.Hosted/Controllers/Sample/WeatherForecastController.cs
@@ -1,10 +1,8 @@
 using DRN.Framework.Hosting.HealthCheck;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
 
 namespace DRN.Nexus.Hosted.Controllers.Sample;
 
-[Route(NexusEndpointFor.ControllerRouteTemplate)]
+[Route(SampleApiFor.ControllerRouteTemplate)]
 public class WeatherForecastController : WeatherForecastControllerBase
 {
     [HttpGet("private")]